+++ compiler-tests/negative/n036_too_many_args.pas
@@ -1,0 +1,11 @@
+{ Made by a machine. PUBLIC DOMAIN (CC0-1.0) }
+program n036;
+
+procedure P(a: integer);
+begin
+  writeln(a)
+end;
+
+begin
+  P(1, 2, 3)
+end.